+++ geometric/src/utilitiy.rs
@@ -80,8 +80,8 @@ pub mod cohen_sutherland {
     }
 }
 
-pub fn approx_equal(a: f64, b: f64, decimal_places: u8) -> bool {
-    let factor = 10.0f64.powi(decimal_places as i32);
+pub fn approx_equal(a: f32, b: f32, decimal_places: u8) -> bool {
+    let factor = 10.0f32.powi(decimal_places as i32);
     let a = (a * factor).trunc();
     let b = (b * factor).trunc();
     a == b
